moves transformer dataset generators to new tool layout
