+++ tools/generators/dataset_generator/main.go
@@ -1,5 +1,3 @@
-//go:build ignore
-
 package main
 
 import (
