+++ tools/tools.go
@@ -1,7 +1,7 @@
 //go:build tools
 // +build tools
 
-package main
+package tools
 
 import (
 	_ "github.com/golang-migrate/migrate/v4/cmd/migrate"
