+++ worker/pkg/benthos/transformers/data-sets/generators.go
@@ -1,7 +1,7 @@
 package transformers_dataset
 
-//go:generate go run data_generator.go datasets/emaildomains.txt EmailDomain $GOPACKAGE
-//go:generate go run data_generator.go datasets/first_names.txt FirstName $GOPACKAGE
-//go:generate go run data_generator.go datasets/last_names.txt LastName $GOPACKAGE
-//go:generate go run business_names.go -output=datasets/business_names.txt
-//go:generate go run data_generator.go datasets/business_names.txt BusinessName $GOPACKAGE
+//go:generate go run ../../tools/generators/dataset_generator/main.go datasets/emaildomains.txt EmailDomain $GOPACKAGE
+//go:generate go run ../../tools/generators/dataset_generator/main.go datasets/first_names.txt FirstName $GOPACKAGE
+//go:generate go run ../../tools/generators/dataset_generator/main.go datasets/last_names.txt LastName $GOPACKAGE
+//go:generate go run ../../tools/generators/business_names/main.go datasets/business_names.txt
+//go:generate go run ../../tools/generators/dataset_generator/main.go datasets/business_names.txt BusinessName $GOPACKAGE
